Refactor which into utils, and find_runtime tweaks
