+++ thx/tests/runner.py
@@ -1,51 +1,19 @@
 # Copyright 2022 Amethyst Reese
 # Licensed under the MIT License
 
-import platform
 import sys
 from asyncio.subprocess import PIPE
 from pathlib import Path
 from unittest import skipIf, TestCase
-from unittest.mock import ANY, call, Mock, patch
+from unittest.mock import ANY, Mock, patch
 
 from .. import runner
 from ..types import CommandError, CommandResult, Config, Context, Job, Result, Version
+from ..utils import venv_bin_path
 from .helper import async_test
 
 
 class RunnerTest(TestCase):
-    @patch("thx.runner.shutil.which")
-    def test_which(self, which_mock: Mock) -> None:
-        context = Context(Version("3.10"), Path(), Path("/fake/venv"))
-        fake_venv_bin = (
-            "/fake/venv/Scripts" if platform.system() == "Windows" else "/fake/venv/bin"
-        )
-        with self.subTest("found"):
-            which_mock.side_effect = lambda b, path: f"/usr/bin/{b}"
-            self.assertEqual("/usr/bin/frobfrob", runner.which("frobfrob", context))
-            which_mock.assert_has_calls([call("frobfrob", path=fake_venv_bin)])
-
-        with self.subTest("not in venv"):
-            which_mock.side_effect = [None, "/usr/bin/scoop"]
-            self.assertEqual("/usr/bin/scoop", runner.which("scoop", context))
-            which_mock.assert_has_calls(
-                [
-                    call("scoop", path=fake_venv_bin),
-                    call("scoop"),
-                ]
-            )
-
-        with self.subTest("not found"):
-            which_mock.side_effect = None
-            which_mock.return_value = None
-            self.assertEqual("frobfrob", runner.which("frobfrob", context))
-            which_mock.assert_has_calls(
-                [
-                    call("frobfrob", path=fake_venv_bin),
-                    call("frobfrob"),
-                ]
-            )
-
     @patch("thx.runner.which")
     def test_render_command(self, which_mock: Mock) -> None:
         which_mock.return_value = "/opt/bin/frobfrob"
@@ -54,7 +22,7 @@ class RunnerTest(TestCase):
         result = runner.render_command("frobfrob check {module}.tests", context, config)
         self.assertEqual(("/opt/bin/frobfrob", "check", "alpha.tests"), result)
 
-    @patch("thx.runner.shutil.which", return_value=None)
+    @patch("thx.utils.shutil.which", return_value=None)
     def test_prepare_job(self, which_mock: Mock) -> None:
         config = Config(values={"module": "beta"})
         context = Context(Version("3.9"), Path(), Path())
@@ -99,7 +67,7 @@ class RunnerTest(TestCase):
                 "/fake/binary", "something", stdout=PIPE, stderr=PIPE, env=ANY
             )
             self.assertIn(
-                str(runner.venv_bin_path(ctx)),
+                str(venv_bin_path(ctx.venv)),
                 exec_mock.call_args.kwargs["env"]["PATH"],
             )
